+++ hawc/apps/udf/apps.py
@@ -3,5 +3,5 @@ from django.apps import AppConfig
 
 class FormLibraryConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "hawc.apps.form_library"
-    verbose_name = "Form Library"
+    name = "hawc.apps.udf"
+    verbose_name = "User Defined Forms"
