+++ hawc/constants.py
@@ -17,7 +17,7 @@ class FeatureFlags(BaseModel):
     ANONYMOUS_ACCOUNT_CREATION: bool = True
     ENABLE_BMDS_33 = False
     ENABLE_PLOTLY_VISUAL: bool = False
-    ENABLE_DYNAMIC_FORMS: bool = False
+    ENABLE_UDF: bool = False
 
     @classmethod
     def from_env(cls, variable) -> "FeatureFlags":
