+++ hawc/main/settings/unittest.py
@@ -10,7 +10,7 @@ DEBUG = True
 
 # enable feature flags for tests
 HAWC_FEATURES.ENABLE_BMDS_33 = True
-HAWC_FEATURES.ENABLE_DYNAMIC_FORMS = True
+HAWC_FEATURES.ENABLE_UDF = True
 
 # remove toolbar for integration tests
 INSTALLED_APPS = [app for app in INSTALLED_APPS if app != "debug_toolbar"]
