+++ hawc/main/urls.py
@@ -10,7 +10,6 @@ import hawc.apps.eco.urls
 import hawc.apps.epi.urls
 import hawc.apps.epimeta.urls
 import hawc.apps.epiv2.urls
-import hawc.apps.form_library.urls
 import hawc.apps.hawc_admin.urls
 import hawc.apps.invitro.urls
 import hawc.apps.lit.urls
@@ -18,6 +17,7 @@ import hawc.apps.mgmt.urls
 import hawc.apps.riskofbias.urls
 import hawc.apps.study.urls
 import hawc.apps.summary.urls
+import hawc.apps.udf.urls
 import hawc.apps.vocab.urls
 from hawc.apps.assessment import views
 from hawc.apps.common.autocomplete import get_autocomplete
@@ -44,7 +44,7 @@ urlpatterns = [
     path("epidemiology/", include("hawc.apps.epiv2.urls")),
     path("epi-meta/", include("hawc.apps.epimeta.urls")),
     path("in-vitro/", include("hawc.apps.invitro.urls")),
-    path("forms/", include("hawc.apps.form_library.urls")),
+    path("udf/", include("hawc.apps.udf.urls")),
     path("bmd/", include("hawc.apps.bmd.urls")),
     path("lit/", include("hawc.apps.lit.urls")),
     path("summary/", include("hawc.apps.summary.urls")),
